requirements.txt
accelerate==0.20.3
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
diffusers==0.17.1
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.6
lpips==0.1.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
wandb==0.15.4
zipp==3.15.0
